mains: Added separate builtin entry and update points for desktop and wasm targets
So, now instead of starting everything from ``pub fn main() void``, the
user has to make three functions, namely:
- ``pub fn init() !void``
- ``pub fn update() !bool``
- ``pub fn deinit() void``

The goal is that build script would use different entry point in
different platforms. So for desktop, it is ``pub fn main()
anyerror!void`` as usual. But for Wasm, the corresponding entry
functions are custom.

In order to achieve this, a build helper (Sdk.zig) is needed.

In practice, this will be optional and configured from build file.