build: Add www/template.html to wasm install options
Also set the default (non change-able) name for any wasm project to
'application'. This is done to simplify the process of testing as
template.html has hardcoded the wasm file name to be 'application.wasm'.
The goal is to be able to in-memory format template.html, but no such
(easy to use) solution exists now.