wasm: Change title of tab when mouse enters a canvas
The behavior is not the most perfect. I actually want the title to
change when the canvas catches input focus. Still need to figure that
out.

Also, while untested, I believe it doesnt work if wasm side calls
setTitle but mouse already is inside the canvas, which ofcouse is a bug.