wasm: Redesign event processing system and fix issue with canvas key events
Moved the TailQueue from zig side to js side. Now on each event
callback, the necessary data (all in integer form) will be pushed to the
event queue on js side and then later shifted from the zig side. This is
unsafe but as long as we, the maintainers, aren't cautious. Overally as
it has been completely implemented, it wont cause any problem and is
completely safe now.

For the canvas key events not working, the solution was to set a tab
index for the canvas element which makes it able to focus from mouse
clicks.